+++ src/render/native/core/utils/utils.hpp
@@ -54,7 +54,7 @@ class MemoryPool {
 			free_block_head = &(new_chunk->blocks[0]);
 
 			// link the new chunk's all blocks
-			for (int i = 1; i < BlockNum; i++)
+			for (size_t i = 1; i < BlockNum; i++)
 				new_chunk->blocks[i - 1].next = &(new_chunk->blocks[i]);
 			new_chunk->blocks[BlockNum - 1].next = NULL; // final block next is NULL
